+++ CHANGELOG.md
@@ -182,3 +182,4 @@
 * Define `UU_U_I_III_U_II_III` data type
 * Define `Diagram` type alias
 * Define `Diagonal` type alias
+* Define `Cone` type class